ddt out of the /qiskit module (#2897)
* from unittest.mock import patch

* .

* move ddt out

* lint
