+++ qiskit/test/__init__.py
@@ -15,7 +15,6 @@
 """Functionality and helpers for testing Qiskit."""
 
 from .base import QiskitTestCase
-from .decorators import requires_aer_provider, online_test, slow_test, requires_qe_access, \
-    combine
+from .decorators import requires_aer_provider, online_test, slow_test, requires_qe_access
 from .reference_circuits import ReferenceCircuits
 from .utils import Path
